Move Cardano.CLI.Json.Friendly to Cardano.CLI.Compatible.Json.Friendly
